@@ -1,19 +1,21 @@
 import React from 'react';
 import store from './Redux/redux-store';
-import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
-import { HashRouter} from 'react-router-dom';
+import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
+import { createRoot } from 'react-dom/client';
 
- 
-  ReactDOM.render(
-   
-      <HashRouter>
-      <Provider store={store}>
+
+
+const container = document.getElementById('root')
+const root = createRoot(container)
+
+root.render(
+
+  <BrowserRouter>
+    <Provider store={store}>
       <App />
-      </Provider>
-      </HashRouter>
-    ,
-    document.getElementById('root'));
+    </Provider>
+  </BrowserRouter>);
 
