@@ -1,5 +1,5 @@
 import { getAuth, getCaptchaURL, getLogin, getLogout } from "../API/api"
-import { getProfileThunk, getStatusThunk } from "./profileReducer"
+import { getProfileThunk, getStatusThunk, setLoading } from "./profileReducer"
 const SET_USER_DATA = 'SET_USER_DATA'
 const SET_LOGIN = 'SET_LOGIN'
 const SET_LOGOUT = 'SET_LOGOUT'
@@ -107,7 +107,10 @@ export const getAuthThunk = () => async (dispatch) => { // санкриэйто�
     if (data.resultCode === 0) {
         dispatch(setAuthUserData(data.data))
         dispatch(getProfileThunk(data.data.id))
-        dispatch(getStatusThunk(data.data.id))
+        dispatch(getStatusThunk(data.data.id))  
+    }
+    else {
+        dispatch(setLoading(false))
     }
 }
 
