@@ -4,7 +4,7 @@ import messagesReduser from "./messagesReducer";
 import profileReduser from "./profileReducer";
 import usersReduser from "./usersReducer";
 import  thunkMiddleware from "redux-thunk"
-import appReduser from "./AppReducer";
+
 
 
 
@@ -14,7 +14,6 @@ let reducers = combineReducers(
         messagesData: messagesReduser,
         usersPage: usersReduser,
         authData: authReduser,
-        app: appReduser,
       
     }
 )
